instance on prepare
Hey,

I am using the command line a lot and every commonds takes 2 seconds.
A similar api call is less than 0.5s. I think the cli can be improved to reduce that time. Nodejs is not that bad after all :)

I have just checked quickly and came accross a piece of code that I do not understand the purpose.
By removing it it reduces the reponse time by 25% (0.5s) and looks to provide the same output in the couple of commands I have tested.

Are those couple lines still valid ?